requirements.txt
python-telegram-bot
instaloader
python-dotenv
yt-dlp
browser-cookie3